examples/910_polyglot_demo/py/requirements.txt
coherence-client==2.0.0
Quart==0.20.0
